refactor: update package name
