--- azra_bills/admin/models/__init__.py
@@ -1,7 +1,0 @@
-from azra_bills.admin.models.city import City as City
-from azra_bills.admin.models.country import Country as Country
-from azra_bills.admin.models.holiday import Holiday as Holiday
-from azra_bills.admin.models.saas_admin import SAASAdmin as SAASAdmin
-from azra_bills.admin.models.state import State as State
-from azra_bills.admin.models.store import Store as Store
-from azra_bills.admin.models.store import StoreContactDetail as StoreContactDetail
