+++ azra_bills_api/admin/models/store.py
@@ -5,11 +5,11 @@ from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.sql import expression
 
-from azra_bills.models import BaseModalWithSoftDelete
+from azra_bills_api.models import BaseModalWithSoftDelete
 from core.utils import ULIDGenerator
 
 if TYPE_CHECKING:
-    from azra_bills.admin.models import City, Country, Holiday, SAASAdmin, State
+    from azra_bills_api.admin.models import City, Country, Holiday, SAASAdmin, State
 
 
 class Store(BaseModalWithSoftDelete):
