+++ azra_bills_api/models.py
@@ -2,7 +2,7 @@ from sqlalchemy import DateTime, func
 from sqlalchemy.ext.asyncio import AsyncAttrs, AsyncSession
 from sqlalchemy.orm import mapped_column
 
-from azra_bills.mixins import SoftDeleteMixin
+from azra_bills_api.mixins import SoftDeleteMixin
 from core.database import Base
