+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "azra_bills"
+name = "azra_bills_api"
 version = "1.0.0"
 description = "A SaaS-based laundry management system for single and multi-store setups. Enables bill generation, updates, customer assignment, and cash logging. Customers can view and pay bills via UPI, wallets, or cards and receive status reminders. Features dashboards, transaction reports, and high-value alerts for streamlined, secure operations."
 authors = ["Tariq Jamal A <tariqjamal4267324@gmail.com>"]
